# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api_key.APIKeyResponse:
    properties:
      created_at:
        type: string
      display_key:
        description: Masked key for display
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used:
        type: string
      max_usage_count:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
    required:
    - created_at
    - display_key
    - id
    - name
    - updated_at
    - usage_count
    type: object
  api_key.APIKeyWithTokenResponse:
    properties:
      created_at:
        type: string
      display_key:
        description: Masked key for display
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used:
        type: string
      max_usage_count:
        type: integer
      name:
        type: string
      token:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
    required:
    - created_at
    - display_key
    - id
    - name
    - token
    - updated_at
    - usage_count
    type: object
  api_key.CreateAPIKeyDto:
    properties:
      expires_at:
        type: string
      max_usage_count:
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  api_key.UpdateAPIKeyDto:
    properties:
      expires_at:
        type: string
      max_usage_count:
        minimum: 1
        type: integer
      name:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
  utils.APIError:
    properties:
      message:
        type: string
      status:
        type: string
    required:
    - message
    - status
    type: object
...
  utils.ApiResponse-api_key_APIKeyResponse:
    properties:
      data:
        $ref: '#/definitions/api_key.APIKeyResponse'
      message:
        type: string
    required:
    - data
    - message
    type: object
  utils.ApiResponse-api_key_APIKeyWithTokenResponse:
    properties:
      data:
        $ref: '#/definitions/api_key.APIKeyWithTokenResponse'
      message:
        type: string
    required:
    - data
    - message
    type: object
  utils.ApiResponse-array_api_key_APIKeyResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api_key.APIKeyResponse'
        type: array
      message:
        type: string
    required:
    - data
    - message
    type: object
...
info:
  contact: {}
  title: Vigi API
paths:
  /api-keys:
    get:
      description: Get all API keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-array_api_key_APIKeyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Get API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key
      parameters:
      - description: API key creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_key.CreateAPIKeyDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse-api_key_APIKeyWithTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Create API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      description: Delete an API key
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Delete API key
      tags:
      - api-keys
    get:
      description: Get a specific API key by ID
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-api_key_APIKeyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Get API key
      tags:
      - api-keys
    put:
      consumes:
      - application/json
      description: Update an API key
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      - description: API key update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_key.UpdateAPIKeyDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-api_key_APIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Update API key
      tags:
      - api-keys
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication (pk_ prefix format)
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    description: JWT token authentication (Bearer token format)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"